refactor: move go/{timer, netutil, grpccommon} under go/tools
The ovearll direction is:
- go/tools has no dependencies on the rest of the repo
- go/common can depend on go/pb and go/tools
- go/services/* can depend on common, pb, and tools, but a service cannot depend on other service code. Shared logic (like a grpc client wrapper) would go under go/common
- go/cmd can depend on anything
