+++ go/tools/netutil/conn.go
