+++ go/tools/netutil/netutil.go
