+++ go/tools/timer/timer.go
