+++ go/tools/timer/timer_test.go
